+++ CHANGELOG.md
@@ -33,6 +33,7 @@ ENHANCEMENT:
 - sequentium.sequence.sequences.integer.explicit.py: added sequence A051682 (hendecagonal numbers)
 - sequentium.sequence.sequences.integer.explicit.py: added sequence A051624 (dodecagonal numbers)
 - sequentium.sequence.sequences.integer.explicit.py: added sequence A051865 (tridecagonal numbers)
+- sequentium.sequence.sequences.integer.explicit.py: added sequence A051866 (tetradecagonal numbers)
 
 ---
 ## [0.0.1] - 2024-01-02