+++ CHANGELOG.md
@@ -35,6 +35,7 @@ ENHANCEMENT:
 - sequentium.sequence.sequences.integer.explicit.py: added sequence A051865 (tridecagonal numbers)
 - sequentium.sequence.sequences.integer.explicit.py: added sequence A051866 (tetradecagonal numbers)
 - sequentium.sequence.sequences.integer.explicit.py: added sequence A051867 (pentadecagonal numbers)
+- sequentium.sequence.sequences.integer.explicit.py: added sequence A051868 (hexadecagonal numbers)
 
 ---
 ## [0.0.1] - 2024-01-02