+++ CHANGELOG.md
@@ -39,6 +39,7 @@ ENHANCEMENT:
 - sequentium.sequence.sequences.integer.explicit.py: added sequence A051869 (heptadecagonal numbers)
 - sequentium.sequence.sequences.integer.explicit.py: added sequence A051870 (octadecagonal numbers)
 - sequentium.sequence.sequences.integer.explicit.py: added sequence A051871 (enneadecagonal numbers)
+- sequentium.sequence.sequences.integer.explicit.py: added sequence A051872 (icosagonal numbers)
 
 ---
 ## [0.0.1] - 2024-01-02