+++ CHANGELOG.md
@@ -27,6 +27,7 @@ ENHANCEMENT:
 - sequentium.sequence.sequences.integer.periodic_generalised.py: added class for representing constant sequences
 - sequentium.sequence.sequences.integer.recursive.py: added sequence A001608
 - sequentium.sequence.sequences.integer.explicit.py: added sequence A000566 (heptagonal numbers)
+- sequentium.sequence.sequences.integer.explicit.py: added sequence A000567 (octagonal numbers)
 
 ---
 ## [0.0.1] - 2024-01-02