+++ CHANGELOG.md
@@ -32,6 +32,7 @@ ENHANCEMENT:
 - sequentium.sequence.sequences.integer.explicit.py: added sequence A001107 (decagonal numbers) 
 - sequentium.sequence.sequences.integer.explicit.py: added sequence A051682 (hendecagonal numbers)
 - sequentium.sequence.sequences.integer.explicit.py: added sequence A051624 (dodecagonal numbers)
+- sequentium.sequence.sequences.integer.explicit.py: added sequence A051865 (tridecagonal numbers)
 
 ---
 ## [0.0.1] - 2024-01-02