+++ CHANGELOG.md
@@ -37,6 +37,7 @@ ENHANCEMENT:
 - sequentium.sequence.sequences.integer.explicit.py: added sequence A051867 (pentadecagonal numbers)
 - sequentium.sequence.sequences.integer.explicit.py: added sequence A051868 (hexadecagonal numbers)
 - sequentium.sequence.sequences.integer.explicit.py: added sequence A051869 (heptadecagonal numbers)
+- sequentium.sequence.sequences.integer.explicit.py: added sequence A051870 (octadecagonal numbers)
 
 ---
 ## [0.0.1] - 2024-01-02